pyproject.toml
[project]
name = "emgtrackpad"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "einops>=0.8.1",
    "h5py>=3.15.1",
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.7",
    "mindrove>=5.2.4",
    "numpy>=2.3.5",
    "omegaconf>=2.3.0",
    "pandas>=2.3.3",
    "pyobjc-framework-quartz>=12.1",
    "rich>=14.0.0",
    "scipy>=1.16.3",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.11",
]
